fixed errors on file import
