+++ src/reducers/index.js
@@ -3,7 +3,6 @@ import initialReducer from './initialReducer';
 import { authReducer } from './authReducer';
 import uiReducer from './uiReducer';
 import profileReducer from './profileReducer';
-// import articleReducer from './articleReducer';
 import imageReducer from './imageReducer';
 
 const reducers = combineReducers({
